semver: Allow parsing of partial version string
Add the `try_parse` function allowing `semver` to parse partial
version string.

When only a single digit is given, it is assumed to be the major
version: `1` is equivalent to `1.0.0`. When two digits are given, it
is assumed to be the major and minor parts of the version string:
`1.2` is equivalent to `1.2.0`.

Signed-off-by: Jimmy Thrasibule <jimmy@thrasibule.mx>