Doc: Add version rule to avoid surpises
Add a new section "Release Policy" to give a
recommendation.

As with the upcoming major 3 release of semver, a lot of
things have been changed. To avoid any surprises for
our users or even break their code base, a short
version restriction can solve that.

This should help users who wants to stay with release 2.

Basically, it boils down to a line like `semver>=2,<3`.
This can be added in any file which lists dependencies.