Fix #258: Keep semver._deprecated.compare
Although it breaks consistency with module level functions, it
seems it's a much needed/used function. Nevertheless it's
a "deprecated" function.

* Function is also available accessing semver.compare()
* Decorate semver.compare() as PendingDeprecationWarning
* Update docstring of semver.compare()
* Adapt `deprecated` decorator and use enforce keyword arguments
* Update CHANGELOG.rst
* Use intersphinx to link to Python exception, use Python
  inventory

It's still unclear if we should deprecate this function or not (that's
why we use PendingDeprecationWarning).
As we don't have a uniform initializer yet, this function stays
in _deprecated.py for the time being until we find a better soltuion.

See #258 for details