se #doctest: +SKIP to avoid OrderedDict result
The OrderedDict.__repr__ method was changed in Python 3.12
to output a different string than before (see #418 for details).

The respective doctests which show the output was marked as
"skipped" with the above directive. We keep the test for documentation
purpuses.

In the test tests/test_semver.py::test_should_versioninfo_to_dict
we test the keys and values separately.